Fix Windows warnings.

+++ test/integration/interaction/BigHelloWorld.c
@@ -45,9 +45,9 @@ bool BigHelloWorld_deserialize_topic(ucdrBuffer* reader, BigHelloWorld* topic)
 uint32_t BigHelloWorld_size_of_topic(const BigHelloWorld* topic, uint32_t size)
 {
     uint32_t previousSize = size;
-    size += ucdr_alignment(size, 4) + 4;
+    size += (uint32_t)(ucdr_alignment(size, 4) + 4);
 
-    size += ucdr_alignment(size, 4) + 4 + (uint32_t)strlen(topic->message) + 1;
+    size += (uint32_t)(ucdr_alignment(size, 4) + 4 + strlen(topic->message) + 1);
 
     return size - previousSize;
 }
